Migration should also check permission exist before rolling back

+++ src/Database/Migrations/v100/Permissions.php
@@ -49,7 +49,9 @@ class Permissions extends Migration
     {
         foreach ($this->getPermissions() as $permission) {
             $permission = Permission::where($permission)->first();
-            $permission->delete();
+            if ($permission) {
+                $permission->delete();
+            }
         }
     }
